fix: build issues/api endpont causing issues
