--- config.overrides.ts
@@ -1,32 +1,0 @@
-import webpack from "webpack";
-
-module.exports = function override(config: any) {
-  const fallback = config.resolve.fallback || {};
-  Object.assign(fallback, {
-    crypto: false,
-    stream: false,
-    assert: false,
-    http: false,
-    https: false,
-    os: false,
-    url: false,
-    zlib: false,
-  });
-  config.resolve.fallback = fallback;
-  config.plugins = (config.plugins || []).concat([
-    new webpack.ProvidePlugin({
-      process: "process/browser",
-      Buffer: ["buffer", "Buffer"],
-    }),
-  ]);
-  config.ignoreWarnings = [/Failed to parse source map/];
-  config.module.rules.push({
-    test: /\.(js|mjs|jsx)$/,
-    enforce: "pre",
-    loader: require.resolve("source-map-loader"),
-    resolve: {
-      fullySpecified: false,
-    },
-  });
-  return config;
-};
