+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@headlessui/react": "1.7.17",
     "@heroicons/react": "2.0.18",
+    "@walletconnect/sign-client": "^2.10.4",
     "@web3auth/base": "7.1.0",
     "@web3auth/metamask-adapter": "7.1.0",
     "@web3auth/modal": "7.1.1",
@@ -23,26 +24,15 @@
     "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "typescript": "5.2.2",
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/elliptic": "^6.4.14",
+    "@types/node": "20.4.1",
+    "@types/react": "18.2.14",
+    "@types/react-dom": "18.2.7",
+    "typescript": "5.1.6",
     "autoprefixer": "^10",
     "postcss": "^8",
     "tailwindcss": "^3",
     "eslint": "^8",
-    "eslint-config-next": "13.5.6",
-    "assert": "^2.0.0",
-    "buffer": "^6.0.3",
-    "crypto-browserify": "^3.12.0",
-    "eslint-config-react-app": "^7.0.1",
-    "https-browserify": "^1.0.0",
-    "os-browserify": "^0.3.0",
-    "process": "^0.11.10",
-    "source-map-loader": "^4.0.1",
-    "stream-browserify": "^3.0.0",
-    "stream-http": "^3.2.0",
-    "url": "^0.11.0",
-    "zlib-browserify": "^0.0.3"
+    "eslint-config-next": "13.5.6"
   }
 }
